requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0
datasets>=2.15.0
matplotlib>=3.8.0
pytest>=7.4.0
pyyaml>=6.0.0
